// frontend/src/lib/Type.ts
interface Title {
  type: "text";
  text: TextContent;
  annotations: Annotations;
  plain_text: string;
  href: string | null;
}

interface DateProperty {
  start: string;
  end: string | null;
  time_zone: string | null;
}

interface MultiSelectOption {
  id: string;
  name: string;
  color: string;
}

interface Properties {
  Date: {
    id: string;
    type: "date";
    date: DateProperty;
  };
  category: {
    id: string;
    type: "multi_select";
    multi_select: MultiSelectOption[];
  };
  title: {
    id: string;
    type: "title";
    title: Title[];
  };
  Cover: {
    id: string;
    type: string;
    files: [
      {
        name: string;
        type: string;
        file: {
          url: string;
          expiry_time: string;
        };
      }
    ];
  };
}

interface User {
  object: "user";
  id: string;
}

interface Parent {
  type: "database_id";
  database_id: string;
}

export interface Page {
  object: "page";
  id: string;
  created_time: string;
  last_edited_time: string;
  created_by: User;
  last_edited_by: User;
  cover: string | null;
  icon: string | null;
  parent: Parent;
  archived: boolean;
  in_trash: boolean;
  properties: Properties;
  url: string;
  public_url: string | null;
}

export interface NotionApi {
  object: "list";
  results: Page[];
  next_cursor: string | null;
  has_more: boolean;
  type: "page_or_database";
  page_or_database: Record<string, unknown>;
  request_id: string;
}